requirements.txt
# ==============================================
# MEDICAL TELEGRAM WAREHOUSE - REQUIREMENTS
# ==============================================

# Core Python Dependencies
psycopg2-binary==2.9.9
python-dotenv==1.0.0
telethon==1.34.0
pandas==2.1.4
requests==2.31.0
python-dateutil==2.8.2

# dbt Core
dbt-core==1.5.0
dbt-postgres==1.5.0

# dbt Packages (installed via packages.yml, listed here for documentation)
# dbt-labs/dbt_utils==1.1.1
# dbt-labs/codegen==0.12.0

# Development tools
pytest==7.4.3
black==23.11.0
flake8==6.1.0
pre-commit==3.5.0
